Question F and H
f)	Provide an example code snippet, in any programming language, that includes the class name, function name, and any other necessary variables, to demonstrate how to achieve the goals of this project. You do not need to include the implementation details of the functions. Please show class diagrams or template code.